Move notification list to component.
