+++ ui/src/app/base/components/NotificationList/index.js
@@ -1,0 +1,1 @@
+export { default } from "./NotificationList";
