+++ ui/src/app/base/components/Section/Section.js
@@ -1,30 +1,12 @@
-import { useDispatch, useSelector } from "react-redux";
 import PropTypes from "prop-types";
 import React from "react";
 
 import "./Section.scss";
-import { messages as messageActions } from "app/base/actions";
 import Col from "app/base/components/Col";
-import Notification from "app/base/components/Notification";
-import selectors from "app/base/selectors";
+import NotificationList from "app/base/components/NotificationList";
 import Strip from "app/base/components/Strip";
 
-const generateMessages = (messages, dispatch) =>
-  messages.map(({ id, message, status, temporary, type }) => (
-    <Notification
-      close={() => dispatch(messageActions.remove(id))}
-      key={id}
-      status={status}
-      timeout={temporary && 5000}
-      type={type}
-    >
-      {message}
-    </Notification>
-  ));
-
 const Section = ({ children, sidebar, title }) => {
-  const messages = useSelector(selectors.messages.all);
-  const dispatch = useDispatch();
   return (
     <div className="section">
       <Strip element="header" className="section__header">
@@ -42,7 +24,7 @@ const Section = ({ children, sidebar, title }) => {
           </Col>
         )}
         <Col size={sidebar ? 10 : 12} className="section__content">
-          {generateMessages(messages, dispatch)}
+          <NotificationList />
           {children}
         </Col>
       </Strip>
